Align language of resource type and name
We define the characteristics of a K8s Resource where
`resource_type` is the type of the resource (e.g `service_account`)
and `resource_name` is its name (e.g `my-service-account`). This commit
aligns a part of the code that didn't implement the definition above and
can confuse the reader.