Use Value Object for K8s Application Identity
Use a Value Object for the k8s resources instead of using
a hash object.

This commit also restructures the `call` function of
`validate_application_identity` to better tell the story
and to align between this class and the corresponding one
of authn-azure.

Application identity configuration validations are now performed
in `validate_application_identity` to improve readability. The
container validations are perform in `validate_pod_request` as the
container name is not part of the application identity.